refactor: restructure directory structure to make it less noisy
