+++ src/components/Canary/ChecksListing.tsx
@@ -5,10 +5,10 @@ import { useGetCheckDetails } from "../../api/query-hooks/health";
 import { EvidenceType } from "../../api/types/evidence";
 import { HealthCheck } from "../../api/types/health";
 import { isCanaryUI } from "../../context/Environment";
+import { Loading } from "../../ui/Loading";
 import { Modal } from "../../ui/Modal";
-import AttachAsEvidenceButton from "../AttachEvidenceDialog/AttachAsEvidenceDialogButton";
 import { timeRanges } from "../Dropdown/TimeRange";
-import { Loading } from "../Loading";
+import AttachAsEvidenceButton from "../Incidents/AttachEvidenceDialog/AttachAsEvidenceDialogButton";
 import PlaybooksDropdownMenu from "../Playbooks/Runs/Submit/PlaybooksDropdownMenu";
 import { CanaryCards } from "./CanaryCards";
 import { CheckDetails } from "./CanaryPopup/CheckDetails";
