+++ src/components/Configs/Changes/ConfigDetailsChanges/ConfigDetailsChanges.tsx
@@ -2,16 +2,16 @@ import { useGetConfigChangesById } from "@flanksource-ui/api/query-hooks/useGetC
 import { ConfigChange } from "@flanksource-ui/api/types/configs";
 import { EvidenceType } from "@flanksource-ui/api/types/evidence";
 import { User } from "@flanksource-ui/api/types/users";
-import AttachAsEvidenceButton from "@flanksource-ui/components/AttachEvidenceDialog/AttachAsEvidenceDialogButton";
 import { Icon } from "@flanksource-ui/components/Icon";
 import { ChangeIcon } from "@flanksource-ui/components/Icon/ChangeIcon";
 import { ConfigIcon } from "@flanksource-ui/components/Icon/ConfigIcon";
-import { Loading } from "@flanksource-ui/components/Loading";
+import AttachAsEvidenceButton from "@flanksource-ui/components/Incidents/AttachEvidenceDialog/AttachAsEvidenceDialogButton";
 import { ViewType } from "@flanksource-ui/types";
 import { Age } from "@flanksource-ui/ui/Age";
 import { Avatar } from "@flanksource-ui/ui/Avatar";
 import { DiffRenderer } from "@flanksource-ui/ui/DiffRenderer/DiffRenderer";
 import { JSONViewer } from "@flanksource-ui/ui/JSONViewer";
+import { Loading } from "@flanksource-ui/ui/Loading";
 import { Modal } from "@flanksource-ui/ui/Modal";
 import ModalTitleListItems from "@flanksource-ui/ui/Modal/ModalTitleListItems";
 import { Stat } from "@flanksource-ui/ui/stats/Stat";
