+++ src/components/Configs/ConfigDetailsTabs.tsx
@@ -3,8 +3,8 @@ import { useAtom } from "jotai";
 import { ReactNode } from "react";
 import { useParams } from "react-router-dom";
 import { useGetConfigByIdQuery } from "../../api/query-hooks";
+import { ConfigsDetailsBreadcrumbNav } from "../../ui/BreadcrumbNav/ConfigsDetailsBreadCrumb";
 import { refreshButtonClickedTrigger } from "../../ui/SlidingSideBar/SlidingSideBar";
-import { ConfigsDetailsBreadcrumbNav } from "../BreadcrumbNav/ConfigsDetailsBreadCrumb";
 import { ErrorBoundary } from "../ErrorBoundary";
 import { Head } from "../Head/Head";
 import { SearchLayout } from "../Layout";
