+++ src/components/Configs/Insights/ConfigAnalysisLink/ConfigInsightsDetailsModal.tsx
@@ -7,8 +7,8 @@ import { Modal } from "../../../../ui/Modal";
 import ModalTitleListItems from "../../../../ui/Modal/ModalTitleListItems";
 import TextSkeletonLoader from "../../../../ui/SkeletonLoader/TextSkeletonLoader";
 import { formatISODate, isValidDate } from "../../../../utils/date";
-import AttachAsEvidenceButton from "../../../AttachEvidenceDialog/AttachAsEvidenceDialogButton";
 import { DescriptionCard } from "../../../DescriptionCard";
+import AttachAsEvidenceButton from "../../../Incidents/AttachEvidenceDialog/AttachAsEvidenceDialogButton";
 import ConfigLink from "../../ConfigLink/ConfigLink";
 import ConfigInsightsIcon from "../ConfigInsightsIcon";
