+++ src/components/Incidents/Changelog/IncidentChangelog.tsx
@@ -6,7 +6,7 @@ import { useIncidentState } from "../../../store/incident.state";
 import { CountBadge } from "../../../ui/Badge/CountBadge";
 import { ClickableSvg } from "../../../ui/ClickableSvg/ClickableSvg";
 import CollapsiblePanel from "../../../ui/CollapsiblePanel/CollapsiblePanel";
-import { Loading } from "../../Loading";
+import { Loading } from "../../../ui/Loading";
 import Title from "../../Title/title";
 import { VerticalSCrollView } from "../../VerticalScrollView/VerticalScrollView";
 import IncidentChangelogItem from "./IncidentChangelogItems";
