+++ src/components/Logs/FilterLogs/LogsSelector.tsx
@@ -1,8 +1,7 @@
-import { useSearchParams } from "react-router-dom";
-import { ReactSelectDropdown } from "../ReactSelectDropdown";
-import { getTopology } from "../../api/services/topology";
 import { useQuery } from "@tanstack/react-query";
-import { StateOption } from "../ReactSelectDropdown";
+import { useSearchParams } from "react-router-dom";
+import { getTopology } from "../../../api/services/topology";
+import { ReactSelectDropdown, StateOption } from "../../ReactSelectDropdown";
 
 type Props = {
   classNames?: string;
