+++ src/components/Topology/Sidebar/TopologyActionBar.tsx
@@ -9,8 +9,8 @@ import { EvidenceType } from "../../../api/types/evidence";
 import { Topology } from "../../../api/types/topology";
 import { useFeatureFlagsContext } from "../../../context/FeatureFlagsContext";
 import { features } from "../../../services/permissions/features";
-import { ActionLink } from "../../ActionLink/ActionLink";
-import { AttachEvidenceDialog } from "../../AttachEvidenceDialog";
+import { ActionLink } from "../../../ui/ActionLink/ActionLink";
+import { AttachEvidenceDialog } from "../../Incidents/AttachEvidenceDialog";
 import PlaybooksDropdownMenu from "../../Playbooks/Runs/Submit/PlaybooksDropdownMenu";
 import TopologySnapshotModal from "../TopologyCard/TopologySnapshotModal";
 import { TopologyConfigLinkModal } from "../TopologyConfigLinkModal/TopologyConfigLinkModal";
