+++ src/components/Users/UserList/index.tsx
@@ -10,7 +10,7 @@ import { CellContext } from "@tanstack/table-core";
 import clsx from "clsx";
 import { useEffect, useMemo, useRef, useState } from "react";
 import { BsTrash } from "react-icons/bs";
-import { withAccessCheck } from "../../AccessCheck/AccessCheck";
+import { withAccessCheck } from "../../Authentication/AccessCheck/AccessCheck";
 import { IconButton } from "../../IconButton";
 
 type UserListProps = {
