+++ src/pages/config/details/ConfigDetailsPage.tsx
@@ -1,8 +1,8 @@
 import { useGetConfigByIdQuery } from "@flanksource-ui/api/query-hooks";
 import { ConfigDetailsTabs } from "@flanksource-ui/components/Configs/ConfigDetailsTabs";
-import { Loading } from "@flanksource-ui/components/Loading";
 import { usePartialUpdateSearchParams } from "@flanksource-ui/hooks/usePartialUpdateSearchParams";
 import { JSONViewer } from "@flanksource-ui/ui/JSONViewer";
+import { Loading } from "@flanksource-ui/ui/Loading";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { useParams } from "react-router-dom";
